Workaround for menu clipping
In my testing, `max-height` seems to be the cause for the right-hand menu clipping.  Disabling the attribute hides the menu, so it is apparently important for some reason.

The current height value is around 1200px in my Chrome, and setting max-height to 2000 did not seem to have any adversary effects.